api_readstash/requirements/base.txt
fastapi==0.110.2
fastapi_resource_server==0.1.2
uvicorn==0.29.0
uvicorn[standard]
python-dotenv==0.21.1
backoff==2.2.1
pydantic-settings==2.2.1
pydantic[email]
orjson==3.10.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
alembic==1.13.1
pytz==2023.3.post1
starlette_exporter==0.17.1
httpx==0.27.0
requests==2.31.0
asyncpg==0.29.0
python-multipart==0.0.9
httpx==0.27.0
redis==5.0.4